requirements.txt
Package                Version
---------------------- -----------
asttokens              2.4.1
colorama               0.4.6
comm                   0.2.2
contourpy              1.2.1
cycler                 0.12.1
debugpy                1.8.1
decorator              5.1.1
et-xmlfile             1.1.0
executing              2.0.1
findspark              2.0.1
fonttools              4.50.0
greenlet               3.0.3
ipykernel              6.29.4
ipython                8.23.0
jedi                   0.19.1
jupyter_client         8.6.1
jupyter_core           5.7.2
kiwisolver             1.4.5
logging                0.4.9.6
matplotlib             3.8.4
matplotlib-inline      0.1.6
mysql-connector-python 8.3.0
nest-asyncio           1.6.0
numpy                  1.26.4
openpyxl               3.1.2
packaging              24.0
pandas                 2.2.1
parso                  0.8.3
pillow                 10.3.0
pip                    24.0
platformdirs           4.2.0
prompt-toolkit         3.0.43
psutil                 5.9.8
pure-eval              0.2.2
py4j                   0.10.9.7
Pygments               2.17.2
PyMySQL                1.1.0
pyparsing              3.1.2
pyspark                3.5.1
python-dateutil        2.9.0.post0
pytz                   2024.1
pywin32                306
pyzmq                  25.1.2
setuptools             65.5.0
six                    1.16.0
SQLAlchemy             2.0.29
stack-data             0.6.3
tornado                6.4
traitlets              5.14.2
typing_extensions      4.10.0
tzdata                 2024.1
wcwidth                0.2.13
